+++ src/App.js
@@ -1,34 +1,11 @@
-import React, { useRef, useState } from 'react';
-
+import React from 'react';
 import './App.css';
-import firebase from 'firebase/compat/app';
-import 'firebase/compat/auth';
-import 'firebase/compat/firestore';
-import 'firebase/compat/analytics';
 
 import { useAuthState } from 'react-firebase-hooks/auth';
-import { useCollectionData } from 'react-firebase-hooks/firestore';
-import { getFirestore } from 'firebase/firestore';
-import { getAnalytics } from 'firebase/analytics';
-import { GoogleAuthProvider, getAuth, signInWithRedirect } from 'firebase/auth';
-
-// initialize Firebase
-firebase.initializeApp({
-  apiKey: process.env.FIREBASE_API_KEY,
-  authDomain: process.env.FIREBASE_AUTH_DOMAIN,
-  projectId: process.env.FIREBASE_PROJECT_ID,
-  storageBucket: process.env.FIREBASE_STORAGE_BUCKET,
-  messagingSenderId: process.env.FIREBASE_MSG_SEND_ID,
-  appId: process.env.FIREBASE_APP_ID,
-  measurementId: process.env.FIREBASE_MSRMNT_ID,
-});
-
-console.log(process.env.FIREBASE_API_KEY)
-
-
-const auth = firebase.auth();
-const firestore = firebase.firestore();
-const analytics = firebase.analytics();
+import ChatRoom from './components/ChatRoom/ChatRoom';
+import SignIn from './components/SignIn/SignIn';
+import SignOut from './components/SignOut/SignOut';
+import { auth } from './firebase.config';
 
 function App() {
   const [user] = useAuthState(auth);
@@ -43,90 +20,6 @@ function App() {
       <section>{user ? <ChatRoom /> : <SignIn />}</section>
     </div>
   );
-}
-
-const SignIn = () => {
-  const signInWinGoogle = () => {
-    const provider = new GoogleAuthProvider();
-    signInWithRedirect(auth, provider);
-  };
-
-  return <button onClick={signInWinGoogle}>Sign in with Google</button>;
 };
-
-const SignOut = () => {
-  return (
-    auth.currentUser && <button onClick={() => auth.signOut()}>Sign Out</button>
-  );
-};
-
-const ChatRoom = () => {
-  const dummy = useRef();
-
-  const messageRef = firestore.collection('message');
-  const query = messageRef.orderBy('createdAt').limit(25);
-
-  const [message] = useCollectionData(query, { idField: 'id' });
-
-  const [formValue, setFormValue] = useState('');
-
-  const sendMessage = async (e) => {
-    e.preventDefault();
-
-    const { uid, photoURL } = auth.currentUser;
-
-    await messageRef.add({
-      text: formValue,
-      createdAt: firebase.firestore.FieldValue.serverTimestamp(),
-      uid,
-      photoURL,
-    });
-
-    setFormValue('');
-    dummy.current.scrollIntoView({ behavior: 'smooth' });
-  };
-
-  return (
-    <>
-      <main>
-        {message &&
-          message.map((msg) => <ChatMessage key={msg.id} message={msg} />)}
-
-        <span ref={dummy}></span>
-      </main>
-
-      <form onSubmit={sendMessage}>
-        <input
-          value={formValue}
-          onChange={(e) => setFormValue(e.target.value)}
-          placeholder="say something nice"
-        />
-
-        <button type="submit" disabled={!formValue}>
-          🕊️
-        </button>
-      </form>
-    </>
-  );
-};
-
-function ChatMessage(props) {
-  const { text, uid, photoURL } = props.message;
-
-  const messageClass = uid === auth.currentUser.uid ? 'sent' : 'received';
-
-  return (
-    <>
-      <div className={`message ${messageClass}`}>
-        <img
-          src={
-            photoURL || 'https://api.adorable.io/avatars/23/abott@adorable.png'
-          }
-        />
-        <p>{text}</p>
-      </div>
-    </>
-  );
-}
 
 export default App;
