--- src/ChatRoom/ChatRoom.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const ChatRoom = () => {
-  return (
-    <div>
-      
-    </div>
-  );
-};
-
-export default ChatRoom;
