+++ src/components/ChatMessage/ChatMessage.jsx
@@ -1,0 +1,23 @@
+import React from 'react';
+import { auth } from '../../firebase.config';
+
+function ChatMessage(props) {
+  const { text, uid, photoURL } = props.message;
+
+  const messageClass = uid === auth.currentUser.uid ? 'sent' : 'received';
+
+  return (
+    <>
+      <div className={`message ${messageClass}`}>
+        <img
+          src={
+            photoURL || 'https://api.adorable.io/avatars/23/abott@adorable.png'
+          }
+        />
+        <p>{text}</p>
+      </div>
+    </>
+  );
+}
+
+export default ChatMessage;
