+++ src/components/ChatRoom/ChatRoom.jsx
@@ -1,0 +1,56 @@
+import React, { useRef, useState } from 'react';
+import { auth, firebase, firestore } from '../../firebase.config';
+import { useCollectionData } from 'react-firebase-hooks/firestore';
+import ChatMessage from '../ChatMessage/ChatMessage';
+
+const ChatRoom = () => {
+  const dummy = useRef();
+
+  const messageRef = firestore.collection('message');
+  const query = messageRef.orderBy('createdAt').limit(25);
+
+  const [message] = useCollectionData(query, { idField: 'id' });
+
+  const [formValue, setFormValue] = useState('');
+
+  const sendMessage = async (e) => {
+    e.preventDefault();
+
+    const { uid, photoURL } = auth.currentUser;
+
+    await messageRef.add({
+      text: formValue,
+      createdAt: firebase.firestore.FieldValue.serverTimestamp(),
+      uid,
+      photoURL,
+    });
+
+    setFormValue('');
+    dummy.current.scrollIntoView({ behavior: 'smooth' });
+  };
+
+  return (
+    <>
+      <main>
+        {message &&
+          message.map((msg) => <ChatMessage key={msg.id} message={msg} />)}
+
+        <span ref={dummy}></span>
+      </main>
+
+      <form onSubmit={sendMessage}>
+        <input
+          value={formValue}
+          onChange={(e) => setFormValue(e.target.value)}
+          placeholder="say something nice"
+        />
+
+        <button type="submit" disabled={!formValue}>
+          🕊️
+        </button>
+      </form>
+    </>
+  );
+};
+
+export default ChatRoom;
