+++ src/components/SignOut/SignOut.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { auth } from '../../firebase.config';
+
+const SignOut = () => {
+  return (
+    auth.currentUser && <button onClick={() => auth.signOut()}>Sign Out</button>
+  );
+};
+
+export default SignOut;
